# [FIX] contacts,*: improve contact computed fields