**Problem**:

The following res_partner computed fields are taking a lot of time to compute for contacts with 500K+ children (+2 minutes)
Fields:
    * total_invoiced
    * meeting_count
    * purchase_order_count
    * sale_order_count
    * supplier_invoice_count

The issue in four of thies five fields is comming from read method.
Prefetching using read was a necessary step before updating orm, as mentioned in the task related to the PR it was added in.
task link: https://www.odoo.com/odoo/all-tasks/1919933

**Solution**:
    * Avoid read method in computation of the fields.
    * Speedup sum generator expression of total_invoiced computation by replacing list with set.

**Benchmark**
| # of children | Before | After |
|--------|--------|--------|
| 8500 | 3s | 2.5s |
| 54000 | 30s | 15s |
| 580000 | 3.20m | 34s |

opw-4421866