@malh-odoo 2025-02-19
Hi @ticodoo 
Thank you for your comments.
I added the mentioned changes and I also added in the commit message the reason of using the "all_partners.read(['parent_id'])" in the first place (after asking the owner of the code).
Additionally, upgrading the client's database is not an easy option for them as they are using SH.
I look forward to your feedback.
Thank you